+++ libs/ballot-interpreter-vx/package.json
@@ -41,6 +41,7 @@
   },
   "dependencies": {
     "@votingworks/ballot-encoder": "workspace:*",
+    "@votingworks/qrdetect": "^1.0.1",
     "@votingworks/types": "workspace:*",
     "@votingworks/utils": "workspace:*",
     "canvas": "^2.6.1",
