+++ libs/ballot-interpreter-vx/src/index.ts
@@ -6,3 +6,4 @@ export {
 export * from './types';
 export { otsu } from './utils/otsu';
 export { crop } from './utils/crop';
+export { detect as detectQrCode } from './utils/qrcode';
