+++ libs/ballot-interpreter-vx/test/fixtures/choctaw-2020-09-22-f30480cc99/hard-qr-code-p1.png
