+++ libs/ballot-interpreter-vx/test/fixtures/jsqr-only-qrcode.png
