+++ services/scan/package.json
@@ -42,7 +42,6 @@
     "@votingworks/fixtures": "workspace:*",
     "@votingworks/logging": "workspace:*",
     "@votingworks/plustek-sdk": "workspace:*",
-    "@votingworks/qrdetect": "^1.0.1",
     "@votingworks/types": "workspace:*",
     "@votingworks/utils": "workspace:*",
     "base64-js": "^1.5.1",
@@ -57,11 +56,9 @@
     "fs-extra": "^9.0.1",
     "got": "^11.8.2",
     "jest-diff": "^26.6.2",
-    "jsqr": "^1.3.1",
     "luxon": "^1.27.0",
     "memory-streams": "^0.1.3",
     "multer": "^1.4.2",
-    "node-quirc": "^2.3.0",
     "ora": "^5.2.0",
     "pdfjs-dist": "2.3.200",
     "tmp": "^0.2.1",
@@ -82,7 +79,6 @@
     "@types/luxon": "^1.26.5",
     "@types/multer": "^1.4.3",
     "@types/node": "^12.0.0",
-    "@types/node-quirc": "workspace:*",
     "@types/pdfjs-dist": "^2.1.3",
     "@types/supertest": "^2.0.10",
     "@types/tmp": "^0.2.0",
