+++ services/scan/test/fixtures/choctaw-2020-09-22-f30480cc99/index.ts
@@ -7,5 +7,3 @@ export const ballotPdf = join(__dirname, 'ballot.pdf');
 export const ballot6522Pdf = join(__dirname, 'ballot-6522.pdf');
 export const blankPage1 = join(__dirname, 'blank-p1.png');
 export const blankPage2 = join(__dirname, 'blank-p2.png');
-export const hardQrCodePage1 = join(__dirname, 'hard-qr-code-p1.png');
-export const hardQrCodePage2 = join(__dirname, 'hard-qr-code-p2.png');
